# api/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PTA scraper",
    "description": "This is API to get journal data from pta.trunojoyo.ac.id",
    "version": "1.0.0"
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  },
  "servers": [
    {
      "url": "https://localhost:3000/"
    },
    {
      "url": "http://localhost:3000/"
    }
  ],
  "tags": [
    {
      "name": "journals",
      "description": "Get journal data"
    }
  ],
  "paths": {
    "/journals/{path}": {
      "get": {
        "tags": [
          "journals"
        ],
        "summary": "Get journal by prodi or faculty",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "description": "prodi or faculty where the journals that want to get",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "fh",
                  "fp",
                  "feb",
                  "ft",
                  "fisib",
                  "fkis",
                  "fkip",
                  "ilmu-hukum",
                  "magister-ilmu-hukum",
                  "teknologi-industri-pertanian",
                  "agribisnis",
                  "agroteknologi",
                  "ilmu-kelautan",
                  "manajemen-sumberdaya-perairan",
                  "magister-pengelolaan-sda",
                  "ekonomi-pembangunan",
                  "manajemen",
                  "akuntansi",
                  "d3-akuntansi",
                  "magister-manajemen",
                  "magister-akuntansi",
                  "d3-enterpreneurship",
                  "magister-ilmu-ekonomi",
                  "teknik-industri",
                  "teknik-informatika",
                  "manajemen-informatika",
                  "teknik-multimedia-dan-jaringan",
                  "mekatronika",
                  "teknik-elektro",
                  "teknik-mekatronika",
                  "sosiologi",
                  "ilmu-komunikasi",
                  "psikologi",
                  "sastra-inggris",
                  "ekonomi-syariah",
                  "hukum-bisnis-syariah",
                  "pgsd",
                  "pendidikan-bahasa-indonesia",
                  "pendidikan-informatika",
                  "pendidikan-ipa",
                  "pgpaud"
                ]
              }
            }
          },
          {
            "name": "startPage",
            "description": "start page of data that will scrape",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "required": false
          },
          {
            "name": "endPage",
            "description": "end page of data that will scrape",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "required": false
          },
          {
            "name": "includeAbstract",
            "description": "if true then the abstract will scrape this might make process longer",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Journal"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Journal": {
        "required": [
          "judul",
          "penulis",
          "pembimbing1",
          "pembimbing2",
          "detailLink"
        ],
        "type": "object",
        "properties": {
          "judul": {
            "type": "string",
            "example": "Klasifikasi citra dengan CNN"
          },
          "penulis": {
            "type": "string",
            "example": "Egi Putra"
          },
          "pembimbing1": {
            "type": "string"
          },
          "pembimbing2": {
            "type": "string"
          },
          "detailLink": {
            "type": "string",
            "example": "https://pta.trunojoyo.ac.id/welcome/detail/{id}"
          },
          "abstaksi": {
            "type": "string",
            "example": "klasifikasi citra adalah ..."
          },
          "abstract": {
            "type": "string",
            "example": "Image classification is ..."
          }
        }
      }
    }
  }
}